pyproject.toml
[project]
name = "manta-dev-gh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "httpx>=0.27.2",
    "pydantic-settings>=2.6.1",
    "pygithub>=2.5.0",
    "ruff>=0.7.4",
    "uvicorn>=0.32.0",
]
